mando de parado de valvulas desde estacion principal

+++ src/services/mqttComunication.py
@@ -52,7 +52,7 @@ class MqttComunication():
                     print('Autorizacion Negociada') 
                     print(self.data[1])
                     if  self.num_GroundDivision == int(self.data[1].split(";")[1]):     
-                        self.NewPrescription = int(self.data[1].split(";")[2])      #se guarda el dato del nuevo valor de prescipcion 
+                        self.NewPrescription = float(self.data[1].split(";")[2])      #se guarda el dato del nuevo valor de prescipcion 
                         print(f'New Prescription = {self.NewPrescription}')
                         self.FlagNewIrrigation = True                            #se activa bandera de riego por negociacion
                 pass
